Adds value errors to `read_image`
When an image path that is given does not exist then it does
not convey this information to the user.
Refactored to raise value error when file does not exist

Images read by the OpenCV might not be a proper image or the
format is not proper which might result in it failing.
Refactored to raise value error with the details.